changes made in the comment section in Ln 7 and Ln 10

+++ Python fundamentals/reading numbers.py
@@ -4,10 +4,10 @@
 # into int() and float() types.
 # lets see the examples below.
 
-marks = float(input('Enter marks:'))        #the function int() around input() converts the read value into integer type. 
+marks = float(input('Enter marks:'))        #the function float() around input() converts the read value into float type. 
 print(type(marks))
 
-age = int(input('Enter age:'))              #the function float() around input() converts the read value into float type. 
+age = int(input('Enter age:'))              #the function int() around input() converts the read value into integer type. 
 print(type(age))
 
 # this is a output with print() function, we use print()
